Separate threshold input setting from checking for valid values
Treat thresholdValue as a string unless doing a comparison.
If you're doing a comparison do it as a number so that differnet
variations of leading and ending 0s will still be equal. If they are
compared as strings they wont be seen as equal.

I think this fixes the threshold input bug.

I dont love that the 'type' of the threshold now gets passed around as a
string but I think that might be the only way to show it correctly in
the input box