Limit supported files to .stl
Thats the only file currently supported by the viewer on the frontend.
In the future we will either build in support for the other file types
in the viewer or we will upload the model file and then download it back
as an stl and place that in the viewer.