Call clearModel() in loadModel()
- remove sceneRef.current.clearModel(); from
useEffect hooks in 3dModelViewer.jsx

Calling sceneRef.current.clearModel();
after loadModel() in the useEffect hook
would clear the previous model. But only because
the model that was being loaded in the useEffect
hook hadn't loaded yet. I think if the loading
process finished quickly for some reason,
the model that we wanted to see would end up
being cleared.